Model changes and fr_locations.json

+++ pyjobs_web/pyjobsweb/model/geocomplete.py
@@ -49,4 +49,4 @@ class Geocomplete(elasticsearch_dsl.DocType):
 
     geolocation = elasticsearch_dsl.GeoPoint()
 
-    weight = elasticsearch_dsl.Integer()
+    weight = elasticsearch_dsl.Float()
